restructure and cleanup
