+++ expertlm/models/__init__.py
@@ -1,0 +1,22 @@
+"""
+Models module for HumanityHelper
+Contains various neural network components for the MoE language model
+"""
+
+from expertlm.models.hybridpositionalencoding import HybridPositionalEncoding
+from expertlm.models.moelanguage import MoELanguageModel
+from expertlm.models.hierarchicalmixtureofexperts import HierarchicalMixtureOfExperts
+from expertlm.models.expertlayer import ExpertLayer
+from expertlm.models.domainspecificattention import DomainSpecificAttention
+from expertlm.models.positionalencoding import PositionalEncoding
+from expertlm.models.specializedexpertnetwork import SpecializedExpertNetwork
+
+__all__ = [
+    'HybridPositionalEncoding',
+    'MoELanguageModel',
+    'HierarchicalMixtureOfExperts',
+    'ExpertLayer',
+    'DomainSpecificAttention',
+    'PositionalEncoding',
+    'SpecializedExpertNetwork',
+]
