+++ expertlm/utils/__init__.py
@@ -1,0 +1,21 @@
+"""
+Utilities module for HumanityHelper
+Contains helper functions and classes for model training, evaluation and data handling
+"""
+
+from expertlm.utils.data import load_and_prepare_datasets, DataConfig, CustomDataCollator
+from expertlm.utils.device import setup_device, optimize_cuda_performance, get_device_map, log_device_info
+from expertlm.utils.moe_trainer import ExpertBalancingTrainer
+from expertlm.utils.preformance import PerformanceMonitor
+
+__all__ = [
+    'load_and_prepare_datasets',
+    'DataConfig',
+    'CustomDataCollator',
+    'setup_device',
+    'optimize_cuda_performance',
+    'get_device_map',
+    'log_device_info',
+    'ExpertBalancingTrainer',
+    'PerformanceMonitor',
+]
