+++ expertlm/utils/preformance.py
@@ -1,0 +1,40 @@
+"""
+File: expert_model.py
+Author: Jeffrey Rivero
+Email: jeff@check-ai.com
+Created: 02/20/2025
+Last Modified: 02/24/2025
+Description: Implements a Mixture-of-Experts (MoE) language model architecture.
+             Includes core components like ExpertLayer, PositionalEncoding,
+             MixtureOfExperts, and the main MoELanguageModel class with
+             generation capabilities.
+"""
+
+import logging
+from typing import Dict, Any
+
+
+# Configure logging
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+class PerformanceMonitor:
+    def __init__(self):
+        self.metrics: Dict[str, Any] = {
+            "forward_pass_times": [],
+            "router_times": [],
+            "expert_processing_times": [],
+            "memory_usage": [],
+            "expert_utilization": [],
+        }
+
+    def update(self, metric_name: str, value: Any):
+        self.metrics[metric_name].append(value)
+
+    def get_summary(self) -> Dict[str, float]:
+        return {
+            name: sum(values) / len(values)
+            for name, values in self.metrics.items()
+            if values
+        }
